tests/wpt/tests/html/semantics/grouping-content/the-pre-element/grouping-pre.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the pre element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-pre-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the pre element.</p>

    <div id="log"></div>

    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("pre");
            assert_equals(Object.getPrototypeOf(testElement), HTMLPreElement.prototype, "HTMLPreElement.prototype should be used for pre");
        }, "The prototype for pre is HTMLPreElement.prototype");

    </script>
</body>
</html>
